pyproject.toml
[project]
name = "house-rent-site"
version = "0.1.0"
description = "House Rent Site - Django REST Framework Project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.12",
    "djangorestframework>=3.14.0",
    "djoser>=2.2.2",
    "djangorestframework-simplejwt>=5.3.2",
    "django-filter>=24.1",
    "drf-yasg>=1.21.7",
    "python-decouple>=3.8",
    "pillow>=10.0.0",
    "django-cors-headers>=4.3.1",
]
